+++ NEWS.md
@@ -19,6 +19,7 @@ Version 1.7.4.9000
 * [#682](https://github.com/tidyverse/lubridate/issues/682) Fix quarter extraction with small `fiscal_start`s.
 * [#703](https://github.com/tidyverse/lubridate/issues/703) `leap_year()` works with objects supported by `year()`.
 
+* `c.Period` concatenation doesn't fail with empty components.
 
 Version 1.7.4
 =============